Split annotation scan UI and harden PDF note rendering
